Move flux points estimator
