+++ gammapy/estimators/points/__init__.py
@@ -1,1 +1,3 @@
 from .lightcurve import *
+from .sensitivity import *
+from .core import *
