+++ gammapy/estimators/points/core.py
@@ -15,11 +15,11 @@ from gammapy.maps.axes import flat_if_equal
 from gammapy.utils.scripts import make_path
 from gammapy.utils.pbar import progress_bar
 from gammapy.utils.table import table_from_row_data, table_standardise_units_copy
-from .flux_map import (
+from ..flux_map import (
     FluxMaps,
     DEFAULT_UNIT,
 )
-from .flux import FluxEstimator
+from ..flux import FluxEstimator
 
 
 __all__ = ["FluxPoints", "FluxPointsEstimator"]
